Fix the unittest
Originally the method test_compute_enriched_elements_new_pressure_without_elasticity had a mock_disc as a fourth argument.
As it was useless, i dedided to test it was not called. Bad idea.
Going back to origin.
The paramter is useless for now. Added a pylint pragma to suppress
warning.